# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BandaBazaar hyperlocal grocery delivery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StoreStatus = (typeof StoreStatus)[keyof typeof StoreStatus];

export const StoreStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  blocked: "blocked",
} as const;

export interface Store {
  id: number;
  name: string;
  ownerPhone: string;
  address: string;
  category: string;
  imageUrl?: string | null;
  status: StoreStatus;
  isOpen: boolean;
  deliveryTimeMinutes?: number;
  rating?: number | null;
  totalRatings?: number;
  minOrderAmount: number;
  commissionRate: number;
  createdAt: string;
}

export interface CreateStoreBody {
  name: string;
  ownerPhone: string;
  address: string;
  category: string;
  imageUrl?: string | null;
  minOrderAmount?: number;
  deliveryTimeMinutes?: number;
}

export interface UpdateStoreBody {
  name?: string;
  address?: string;
  category?: string;
  imageUrl?: string | null;
  isOpen?: boolean;
  minOrderAmount?: number;
  deliveryTimeMinutes?: number;
}

export interface StoreStats {
  storeId: number;
  totalOrders: number;
  totalRevenue: number;
  totalCommission: number;
  netEarnings: number;
  pendingOrders: number;
  completedOrders: number;
}

export interface Product {
  id: number;
  storeId: number;
  name: string;
  nameHindi?: string | null;
  price: number;
  unit?: string | null;
  category: string;
  imageUrl?: string | null;
  inStock: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  nameHindi?: string | null;
  price: number;
  unit?: string | null;
  category: string;
  imageUrl?: string | null;
  inStock?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  nameHindi?: string | null;
  price?: number;
  unit?: string | null;
  category?: string;
  imageUrl?: string | null;
  inStock?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  placed: "placed",
  confirmed: "confirmed",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  storeId: number;
  storeName: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  subtotal: number;
  commissionAmount: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  rating?: number | null;
  ratingComment?: string | null;
  createdAt: string;
  updatedAt?: string | null;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export interface CreateOrderItemBody {
  productId: number;
  quantity: number;
}

export interface CreateOrderBody {
  storeId: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: CreateOrderItemBody[];
  paymentMethod: CreateOrderBodyPaymentMethod;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  confirmed: "confirmed",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface RateOrderBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  ratingComment?: string | null;
}

export interface BlockStoreBody {
  blocked: boolean;
}

export interface AdminDashboard {
  totalStores: number;
  pendingStores: number;
  totalOrders: number;
  totalRevenue: number;
  totalCommission: number;
  recentOrders: Order[];
  topStores: Store[];
}

export interface Commission {
  orderId: number;
  storeId: number;
  storeName: string;
  orderAmount: number;
  commissionRate: number;
  commissionAmount: number;
  createdAt: string;
}

export type ListStoresParams = {
  category?: string;
  search?: string;
};

export type ListProductsParams = {
  category?: string;
  search?: string;
};

export type ListOrdersParams = {
  storeId?: number;
  customerPhone?: string;
  status?: string;
};

export type AdminListStoresParams = {
  status?: string;
};

export type ListCommissionsParams = {
  storeId?: number;
};