# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BandaBazaar hyperlocal grocery delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  storeId: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: CreateOrderItemBody[];
  paymentMethod: CreateOrderBodyPaymentMethod;
}