# lib/api-zod/src/generated/types/createOrderItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BandaBazaar hyperlocal grocery delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderItemBody {
  productId: number;
  quantity: number;
}