# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BandaBazaar hyperlocal grocery delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  storeId: number;
  storeName: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  subtotal: number;
  commissionAmount: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  rating?: number | null;
  ratingComment?: string | null;
  createdAt: string;
  updatedAt?: string | null;
}